Develop

+++ src/Controller/CommentController.php
@@ -79,17 +79,59 @@ class CommentController extends AbstractController
         ]);
     }
 
+     /**
+     * @IsGranted({"ROLE_ADMIN","ROLE_USER"})
+     * @Route("/{id}/edit", name="comment_edit", methods={"GET","POST"})
+     */
+    public function edit(Request $request, Comment $comment, EntityManagerInterface $em): Response
+    {
+        //Check if trick is this current simple_user's trick
+        if ($this->getUser()->getRoles() !== 'ROLE_ADMIN' && $comment->getAuthor()->getId() !== $this->getUser()->getId()) {
+            throw $this->createNotFoundException('Vous ne disposez pas des droits nécessaires pour la modification de ce commentaire.');
+        }
+
+        $form = $this->createForm(CommentType::class, $comment);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+
+            $comment->setUpdatedAt(new \DateTime('now'));
+            $em->flush();
+
+            $this->addFlash('success', 'Super! votre commentaire à bien été modifié.');
+
+            if ($this->getUser()->getRoles() !== 'ROLE_ADMIN') {
+                return $this->redirectToRoute('trick_show', ['id'=>$comment->getTrick()->getId()]);
+            }
+
+            return $this->redirectToRoute('trick_index');
+        }
+
+        return $this->render('comment/edit.html.twig', [
+            'comment' => $comment,
+            'form' => $form->createView(),
+        ]);
+    }
+
     /**
-     * @Security("is_granted('ROLE_USER') and user == comment.author or is_granted('ROLE_ADMIN')", message="Vous ne pouvez pas supprimer ce commentaire.")
+     * @IsGranted({"ROLE_ADMIN","ROLE_USER"})
      * @Route("/{id}", name="comment_delete", methods={"DELETE"})
      */
     public function delete(Request $request, Comment $comment): Response
     {
+        if ($this->getUser()->getRoles() !== 'ROLE_ADMIN' && $comment->getAuthor()->getId() !== $this->getUser()->getId()) {
+            throw $this->createNotFoundException('Vous ne disposez pas des droits nécessaires pour la suppression de ce commentaire.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->request->get('_token'))) {
            $this->em->remove($comment);
            $this->em->flush();
         }
 
+        if ($this->getUser()->getRoles() !== 'ROLE_ADMIN') {
+            return $this->redirectToRoute('trick_show', ['id'=>$comment->getTrick()->getId()]);
+        }
+        
         return $this->redirectToRoute('comment_index');
     }
 }
